test(fido2): migrate previewSign tests to canonical Fido2 layer
- Refactor FidoPreviewSignTests to use ExtensionBuilder.WithPreviewSign()
- Delete test-internal BuildPreviewSignRegistrationInput helper
- Migrate 4 byte-level CBOR encoding tests from WebAuthn.UnitTests to Fido2.UnitTests
- Delete redundant WebAuthn PreviewSignCborEncodingTests.cs and PreviewSignCborTests.cs

Fido2 is the canonical FIDO2 resource; tests follow encoder location.
